CHE-2033 Rename Plugin Github package
