+++ plugins/plugin-github/che-plugin-github-provider-github/src/main/java/org/eclipse/che/ide/ext/git/server/github/GithubGitModule.java
@@ -8,7 +8,7 @@
  * Contributors:
  *   Codenvy, S.A. - initial API and implementation
  *******************************************************************************/
-package org.eclipse.che.ide.ext.git.server.nativegit;
+package org.eclipse.che.ide.ext.git.server.github;
 
 import org.eclipse.che.api.git.CredentialsProvider;
 import org.eclipse.che.inject.DynaModule;
